refactor: service layer
